Moved neighbors parameter to constructor
Moved the optional parameter `nieghbors` to the BlockPartition constructor.
We can now call it by using partition(P, BlockPartition(50.,50., neighbors = true)) or partition(P, BlockPartition(50.,50., neighbors = false)) 

The default value for neighbors is false. If neighbors is not explicitly assigned a value, the function will be called assuming neighbors = false.